requirements.txt
flask
flask_cors
requests
beautifulsoup4
numpy
python-whois
joblib
gunicorn